Move NavigationMeshSourceGeometryData(2D/3D) to the 2D/3D subfolders
