+++ scene/3d/navigation_region_3d.cpp
@@ -30,7 +30,7 @@
 
 #include "navigation_region_3d.h"
 
-#include "scene/resources/navigation_mesh_source_geometry_data_3d.h"
+#include "scene/resources/3d/navigation_mesh_source_geometry_data_3d.h"
 #include "servers/navigation_server_3d.h"
 
 RID NavigationRegion3D::get_rid() const {
