+++ scene/resources/2d/navigation_polygon.cpp
